search function and error handling of fetch for listings and profile pages
