+++ index.html
@@ -103,7 +103,7 @@
             <h2 class="text-center m-2 display-6">Most wanted</h2>
             <div class="row col-12 col-sm-10">
               <div class="most-wanted-listings">
-                <p class="loading">Loading listings</p>
+                <p class="loading-text">Loading listings</p>
               </div>
             </div>
             <a
@@ -119,7 +119,7 @@
             <h2 class="text-center m-2 display-6">Newest</h2>
             <div class="row col-12 col-sm-10">
               <div class="newest-listings">
-                <p class="loading">Loading listings</p>
+                <p class="loading-text">Loading listings</p>
               </div>
             </div>
             <a
@@ -136,7 +136,7 @@
             </div>
 
             <div class="ending-soon-listings position-relative">
-              <p class="loading text-center">Loading listings</p>
+              <p class="loading-text text-center">Loading listings</p>
             </div>
           </div>
         </div>
