+++ profile/index.html
@@ -38,7 +38,7 @@
     </header>
     <main>
       <div
-        class="container-fluid w-100 profile-container main-content border rounded-3 d-flex pb-3 bg-white d-flex flex-column"
+        class="container-fluid w-100 profile-container main-content border rounded-3 d-flex pb-3 bg-white d-flex flex-column mt-5"
       >
         <div class="profile-content d-flex flex-column mt-4 w-100">
           <div class="row align-items-center">
