+++ src/js/api/constants.js
@@ -2,4 +2,5 @@ export const API_URL = "https://v2.api.noroff.dev";
 export const API_URL_AUCTION = `${API_URL}/auction`;
 export const API_URL_LISTINGS = `${API_URL_AUCTION}/listings`;
 export const API_URL_PROFILE = `${API_URL_AUCTION}/profiles`;
+export const API_URL_SEARCH = `${API_URL_LISTINGS}/search`;
 export const API_KEY = "7c4a49fd-2a02-404d-9e34-6b59fe99b3f3";
