+++ src/js/templates/index.js
@@ -3,3 +3,4 @@ export * from "./renderAllListings.js";
 export * from "./renderSingleListing.js";
 export * from "./renderMostBidsListings.js";
 export * from "./renderListingsLandingPage.js";
+export * from "./renderListingsBySearch.js";
