+++ src/js/templates/renderListingsLandingPage.js
@@ -9,8 +9,9 @@ export async function renderListingsLandingPage({
   endingSoon = 5,
 } = {}) {
   let page = 1;
+  const errorMsg = document.querySelector(".main-content");
   try {
-    const firstResponse = await displayListings(1);
+    const firstResponse = await displayListings(1, errorMsg);
     console.log(firstResponse);
     const pageCount = firstResponse.meta.pageCount;
     console.log(`Page count: ${pageCount}`);
@@ -18,7 +19,7 @@ export async function renderListingsLandingPage({
     console.log(`Page ${page} data:`, firstResponse);
 
     for (let i = 2; i <= pageCount; i++) {
-      const response = await displayListings(i);
+      const response = await displayListings(i, errorMsg);
       console.log(`Page ${i} data:`, response);
       allListings = allListings.concat(response.data);
       console.log(allListings);
