+++ src/scss/_main.scss
@@ -1,5 +1,6 @@
 main {
   max-width: 1300px;
+  min-width: 60vw;
   min-height: 90vh;
   margin: auto;
   padding: 1rem 1rem;
